requirements.txt
flask
flask-sqlalchemy
flask-migrate
twilio
redis
python-dotenv
streamlit
pandas
requests
pytest
